# PERF-#7435: Use shallow copies in native pandas mode