--- a/modin/config/envvars.py
+++ b/modin/config/envvars.py
@@ -1348,13 +1348,15 @@ class NativePandasTransferThreshold(EnvironmentVariable, type=int):
 
 class NativePandasDeepCopy(EnvironmentVariable, type=bool):
     """
-    Whether the native pandas backend performs a deep copy of the wrapped pandas DataFrame when
-    constructing a Modin frame from a native pandas object, or when creating a native pandas
-    frame from a Modin one via `df.modin.to_pandas()`.
+    Whether to perform deep copies when transferring data with the native pandas backend.
 
-    Disabling this flag produces significant performance improvements by reducing the number of
-    copy operations performed. However, it may create unexpected results if the user mutates
-    the Modin frame or native pandas frame in-place while pandas's copy-on-write option is disabled:
+    Copies occur when constructing a Modin frame from a native pandas object with
+    `pd.DataFrame(pandas.DataFrame([]))`, or when creating a native pandas frame from a Modin one
+    via `df.modin.to_pandas()`.
+
+    Leaving this flag disabled produces significant performance improvements by reducing the number
+    of copy operations performed. However, it may create unexpected results if the user mutates
+    the Modin frame or native pandas frame in-place.
 
     >>> import pandas  # doctest: +SKIP
     >>> import modin.pandas as pd  # doctest: +SKIP