<!--
Thank you for your contribution!
Please review the contributing docs: https://modin.readthedocs.io/en/latest/development/contributing.html
if you have questions about contributing.
-->

## What do these changes do?

<!-- Please give a short brief about these changes. -->

- [x] first commit message and PR title follow format outlined [here](https://modin.readthedocs.io/en/latest/development/contributing.html#commit-message-formatting)
  > **_NOTE:_**  If you edit the PR title to match this format, you need to add another commit (even if it's empty) or amend your last commit for the CI job that checks the PR title to pick up the new PR title.
- [x] passes `flake8 modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] passes `black --check modin/ asv_bench/benchmarks scripts/doc_checker.py`
- [x] signed commit with `git commit -s` <!-- you can amend your commit with a signature via `git commit -amend -s` -->
- [x] Resolves #7435 <!-- issue must be created for each patch -->
- [x] tests added and passing
- [x] module layout described at `docs/development/architecture.rst` is up-to-date <!-- if you have added, renamed or removed files or directories please update the documentation accordingly -->

Uses shallow copies by default in native pandas execution mode, significantly improving performance. This behavior can be changed by setting `NativePandasDeepCopy.enable()`.

Summary of mutation behavior, with and without pandas copy-on-write:
<img width="1319" height="661" alt="Screenshot 2025-07-31 at 12 29 21" src="https://github.com/user-attachments/assets/8094487c-b8a6-4f4f-a0ab-97bdaf799bd2" />

Unscientific sanity check benchmark:
```
import modin.pandas as pd
import numpy as np
from modin.config import Backend
Backend.put("Pandas")

from time import perf_counter

def bm(f):
    start = perf_counter()
    f() 
    print(perf_counter() - start)

df = pd.DataFrame(np.random.randint(0,100,size=(2**22,2**8)))
bm(lambda: repr(df.sort_values(0)))
```

Native pandas w/ CoW: 2.41s
Native pandas w/o CoW: 4.54s
Modin w/ NativePandasDeepCopy set (equivalent to current main): 7.03s
Modin w/ NativePandasDeepCopy disabled (this PR): 2.30s